separate int-only and all-types header files
