+++ tests/chapter_20/all_types/no_coalescing/force_spill_doubles.c
@@ -1,4 +1,4 @@
-#include "../../libraries/util.h"
+#include "../util.h"
 
 int glob = 3;
 double glob2 = 4.0;
