+++ tests/chapter_20/all_types/no_coalescing/store_pointer_in_register.c
@@ -7,7 +7,7 @@
  * Look for: no spills
  */
 
-#include "../../libraries/util.h"
+#include "../util.h"
 
 int glob1 = 1;
 int glob2 = 2;
