+++ tests/chapter_20/all_types/util.h
