+++ tests/chapter_20/int_only/no_coalescing/callee_saved_stack_alignment.c
@@ -5,7 +5,7 @@
  * exit code.
  * */
 
-#include "../../libraries/util.h" // declares check_* and id functions
+#include "../util.h" // declares check_* and id functions
 
 // check alignment of RSP and exit with exit_code if it's misaligned
 // defined in tests/chapter_20/libraries/alignment_check_<PLATFORM>.s
