+++ tests/chapter_20/int_only/no_coalescing/cmp_generates_operands.c
@@ -3,7 +3,7 @@
  * Note: this only works as intended after we've implemented register
  * coalescing.
  * */
-#include "../../libraries/util.h"
+#include "../util.h"
 
 int glob = 10;
 int glob2 = 20;
