+++ tests/chapter_20/int_only/no_coalescing/eax_live_at_exit.c
@@ -3,7 +3,7 @@
  * Note: only works as intended once we've implemented register coalescing.
  * */
 
-#include "../../libraries/util.h"
+#include "../util.h"
 
 int glob = 10;
 int glob2 = 0;
