+++ tests/chapter_20/int_only/no_coalescing/force_spill.c
@@ -7,7 +7,7 @@
  * This test program is generated from templates/chapter_20_templates/force_spill.c.jinja
  * */
 
-#include "../../libraries/util.h"
+#include "../util.h"
 
 int glob_three = 3;
