+++ tests/chapter_20/int_only/no_coalescing/many_pseudos_fewer_conflicts.c
@@ -3,7 +3,7 @@
  * conflict
  */
 
-#include "../../libraries/util.h"
+#include "../util.h"
 
 // define a function that returns a value
 // so we can't just constant fold everything away
