+++ tests/chapter_20/int_only/no_coalescing/optimistic_coloring.c
@@ -3,7 +3,7 @@
  * spill candidates but only spill 5 of them.
  */
 
-#include "../../libraries/util.h"
+#include "../util.h"
 
 int flag = 0;
 int result = 0;
