+++ tests/chapter_20/int_only/no_coalescing/rewrite_regression_test.c
@@ -8,7 +8,7 @@
  * This test program is generated from templates/chapter_20_templates/rewrite_regression_test.c.jinja
  * */
 
-#include "../../libraries/util.h"
+#include "../util.h"
 
 int glob_three = 3;
 int glob_four = 4;
