+++ tests/chapter_20/int_only/no_coalescing/test_spill_metric.c
@@ -2,7 +2,7 @@
  * we spill the one with the lowest spill cost (i.e. fewest uses)
  */
 
-#include "../../libraries/util.h" // declares check_* and id functions
+#include "../util.h" // declares check_* and id functions
 
 int target(void) {
     // Define 6 callee-saved regs that interfere with each other;
