+++ tests/chapter_20/int_only/no_coalescing/test_spill_metric_2.c
@@ -7,7 +7,7 @@
  * resulting in more memory accesses overall.
  */
 
-#include "../../libraries/util.h"
+#include "../util.h"
 
 int target(void) {
