+++ tests/chapter_20/int_only/no_coalescing/trivially_colorable.c
@@ -2,7 +2,7 @@
  * The test script validates that we don't spill any pseudoregisters in target.
  */
 
-#include "../../libraries/util.h"
+#include "../util.h"
 
 int target(int one, int two) {
     // perform a few calculations that make one, two, three, four, and five
