+++ tests/chapter_20/int_only/no_coalescing/unary_uses_operand.c
@@ -4,7 +4,7 @@
  * TODO consider using a common template for this and bin_uses_operand
  * */
 
-#include "../../libraries/util.h"
+#include "../util.h"
 
 int glob = 1;
