+++ tests/chapter_20/int_only/no_coalescing/use_all_hardregs.c
@@ -3,7 +3,7 @@
  * and make sure we assign all of them to hardregs
  * This test program is generated from templates/chapter_20_templates/use_all_hardregs.c.jinja
  * */
-#include "../../libraries/util.h"
+#include "../util.h"
 
 int global_one = 1;  // to prevent constant-folding
