+++ tests/chapter_20/int_only/with_coalescing/briggs_coalesce_hardreg.c
@@ -10,7 +10,7 @@
  * even if we don't perform register coalescing.
  * */
 
-#include "../../libraries/util.h"
+#include "../util.h"
 
 int flag = 1;
