+++ tests/chapter_20/int_only/with_coalescing/briggs_dont_coalesce.c
@@ -4,7 +4,7 @@
  * sure there are no spills.
  * */
 
-#include "../../libraries/util.h"
+#include "../util.h"
 
 int glob = 5;
