+++ tests/chapter_20/int_only/with_coalescing/george_dont_coalesce.c
@@ -6,7 +6,7 @@
  * sure there are no spills.
  * */
 
-#include "../../libraries/util.h"
+#include "../util.h"
 
 int glob = 1;
