+++ tests/chapter_20/int_only/with_coalescing/george_off_by_one.c
@@ -4,7 +4,7 @@
  * adequately covered by other tests.
  * */
 
-#include "../../libraries/util.h"
+#include "../util.h"
 
 int glob = 0;
